Check for Mon PDB post upgrade
