+++ tests/ecosystem/upgrade/test_upgrade.py
@@ -6,6 +6,8 @@ from ocs_ci.framework.testlib import ocs_upgrade, polarion_id
 from ocs_ci.ocs.disruptive_operations import worker_node_shutdown, osd_node_reboot
 from ocs_ci.ocs.ocs_upgrade import run_ocs_upgrade
 from ocs_ci.utility.reporting import get_polarion_id
+from ocs_ci.ocs.cluster import CephCluster
+from ocs_ci.helpers.helpers import get_mon_pdb
 
 log = logging.getLogger(__name__)
 
@@ -63,3 +65,25 @@ def test_upgrade():
     """
 
     run_ocs_upgrade()
+
+
+@ocs_upgrade
+@pytest.mark.polarion_id("OCS-2449")
+def test_check_mon_pdb_post_upgrade():
+    """
+    Function to check disruptions_allowed and minimum
+    available mon count
+
+    """
+    ceph_obj = CephCluster()
+
+    # Check for mon count
+    mons_after_upgrade = ceph_obj.get_mons_from_cluster()
+    log.info(f"Mons after upgrade {mons_after_upgrade}")
+
+    disruptions_allowed, min_available_mon = get_mon_pdb()
+    log.info(f"Number of Mons Disruptions_allowed {disruptions_allowed}")
+    log.info(f"Minimum_available mon count {min_available_mon}")
+
+    assert disruptions_allowed == 1, "Mon Disruptions_allowed count not matching"
+    assert min_available_mon == 2, "Minimum available mon count is not matching"
